requirements.txt
# コア・ヘッドレス専用依存（PyQt6 は含まない）
# 再発防止: GUI ライブラリをここに追加しないこと。GUI 依存は requirements-gui.txt に記載する。
numpy
# 再発防止: ネイティブ(C++) 側の FetchContent 版と不一致になると DLL/API ミスマッチが再発するため固定する。
onnxruntime==1.20.1
pydantic>=2.0
